Update macros to work from "print" file
RRF does not handle M291 commands the same when running them from the
web console and from within a gcode file, especially when they are
nested within conditionals. For this reason, I have had to re-work a
number of the macros that prompt for user input but can also be given
parameter strings (e.g. `G6001 W3` to set WCS origin on system 3
based on a corner probe).

There are a lot of changes here, but the general gist is separating out
the macros that prompt for user input, and then calling an underlying
macro that expects _only_ parameters to be passed and does not prompt
for user input except for jogging controls where necessary.

The underlying macros can be called directly _if you do not need to
prompt the user for any of the arguments_.

Unfortunately, if you need to prompt the user, you need to prompt them
for _all_ of the parameters - there is no way to effectively mix and
match the parameter / prompt approach and have it work across console
and "print" use.

 - `G6000` becomes `G6000` for prompts and `G6000.1` for parameters.
 - `G6001` becomes `G6001` for prompts and `G6001.1` for parameters.
 - `G7000` (enable HSSC) becomes `M7000` as it does not trigger machine
   movement.
 - `G7001` added to disable HSSC, can be executed while spindle is
   running.
 - `M4000` and `M4001` added to add and remove tool details (an index
   and a description). These details can be used during tool changes to
   prompt the user with the name of the tool as well as the index.
 - `G6999` and `G6999.1` added to prompt user to add / remove their
   touch probe after probing events. This works in combination with all
   other touch probe macros to allow them to be run in sequence
   (e.g. probing the reference surface and then probing a work piece
   without being prompted twice to plug your probe in). Users will only
   be prompted to plug their probe in once, and you _must_ call this
   when a user is expected to unplug their probe so that they are
   prompted to plug it back in the next time there is a probing
   requirement!
 - Change Z probing to not move more than 1mm downwards after the
   initial probing move. This should limit damage in instances of probe
   failure.
 - Various bug fixes.

Signed-off-by: Ben Agricola <ben+git@agrico.la>