Implement spindle deactivation on probing
When changing tools or running probing steps during an active "print"
(executing commands from a file), you are always one command away
from the spindle starting up unexpectedly. To avoid this as much as we
can, we now remove the spindle control config from RRF entirely during
tool changes and other types of probing.

This means that as long as your spindle is wired and configured
correctly, RRF has ZERO control over it while your hands are near it,
reducing the chances of a bug in this configuration from causing you
damage.

Running `M7002` (renamed from `G6999`) will now deactivate the spindle
and prompt the user to insert the touch probe. Once the touch probe is
confirmed connected by the operator, you will not be re-prompted to
connect it until you have run `M7003` (new command). The new `M7003`
re-activates the spindle after the operator confirms it is safe to do
so, and the next time a touch probe is needed, the operator will be
prompted to re-connect it, after the spindle is disabled again.

Your post-processor (or you, the operator) are responsible for running
`M7003` after probing is complete. You will need to run it to
re-activate the spindle prior to any cutting moves.

Signed-off-by: Ben Agricola <ben+git@agrico.la>