Track tool offsets by index
As part of the work to allow our spindle to be deactivated during tool
changes and probing moves, we lost the ability to set the offset
correctly (as offsets are wiped when spindles are added). We now track
the probed offset in a global variable, allowing it to be set when the
tool is activated.

  - `Added M917` - Motor Standstill Current Reduction.
    - This is very useful on CNC machines as well as 3D printers, where
      the Z motor will sit at a standstill consuming as much current as
      it can from the driver while the X and Y axes perform a cut. We
      enable standstill current reduction to avoid driver overheats
      during long operations with driver(s) in standstill for a long
      period of time.
  - Made sure homing scripts run using machine co-ordinates, so they
    don't screw up co-ordinates when run from the web interface.

Signed-off-by: Ben Agricola <ben+git@agrico.la>