Reorganise files into new structure
There's enough files in the repository now to make it look a little
overwhelming to the novice user. As the idea for this system is to host
a github site which allows generation of a configuration based on user
selected options, it doesn't really matter whether the files in this
repository are in the "right" places for RepRapFirmware to use them, as
we can move them to the right places when we generate a configuration
download for each user.

We split out files by type:

  - System configuration types (generally calling internal RRF macros
    to perform the on-boot configuration of the machine).
    - Moved to `/system` in the repository.
    - Should be moved to `/sys` in download package.
  - Macros, which deal with functionality post-boot. These are further
    split down into:
    - Movement Macros, which must be `G` commands, and are any command
      that triggers a movement for RRF to implement.
    - Machine Macros, which must be `M` commands, and are any command
      that configures a part of the machine in-operation but does not
      cause any movements to occur.
    - Both of the above should be moved into `/sys` in the download
      package.
    - Tool Macros, which deal with any specific activities necessary to
      configure the tool (Spindle). Are not moved to `/sys` so cannot be
      called under a `G` or `M` code. Should not be called manually by
      the operator either.
    - Misc Macros, which deal with anything not easily grouped into one
      of the previous categories. Should not be called manually by the
      operator.

Signed-off-by: Ben Agricola <ben+git@agrico.la>