Remove all macros
In preparation for the release of "something new", I'm removing all of
the macros from the baseline configuration.

To ease management, I have separated the parts of this that were related
to actual machine operation into a new library, which should be
independent of your actual RRF configuration. This repository will
become a nice, barebones RRF setup with config split into files and
relying on the single-vars-file format for configuration.

The new tool will be a package of macros that can be installed on any
RRF instance in CNC mode regardless of your config setup, and these two
will be able to be used together as well.

Signed-off-by: Ben Agricola <ben+git@agrico.la>